requirements.txt
python-docx
PyPDF2
tiktoken
regex
openpyxl
pandas